+++ backend/integration_tests/src/set_message_reminder_tests.rs
@@ -18,6 +18,14 @@ fn set_message_reminder_succeeds() {
     let now = now_millis(env);
     let notes = random_string();
 
+    let starting_index =
+        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, EventIndex::default(), true, 1000, 1000)
+            .events
+            .last()
+            .unwrap()
+            .index
+            .incr();
+
     client::user::set_message_reminder(
         env,
         user1.principal,
@@ -31,33 +39,38 @@ fn set_message_reminder_succeeds() {
         },
     );
 
-    let latest_bot_message_index =
-        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, EventIndex::default(), true, 1000, 1000)
-            .events
-            .last()
-            .unwrap()
-            .index;
-
     env.advance_time(Duration::from_millis(999));
     env.tick();
 
-    assert!(
-        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, latest_bot_message_index.incr(), true, 1, 1)
-            .events
-            .is_empty()
-    );
+    let events_response1 =
+        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, starting_index, true, 1000, 1000).events;
+
+    assert_eq!(events_response1.len(), 1);
 
     env.advance_time(Duration::from_millis(1));
     env.tick();
 
-    let latest_bot_message_response =
-        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, latest_bot_message_index.incr(), true, 1, 1);
+    let latest_bot_messages_response =
+        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, starting_index, true, 1000, 1000);
 
-    assert!(!latest_bot_message_response.events.is_empty());
+    assert_eq!(latest_bot_messages_response.events.len(), 2);
 
-    let latest_bot_message = latest_bot_message_response.events.into_iter().next().unwrap();
+    let mut iter = latest_bot_messages_response.events.into_iter();
+    let reminder_created_event = iter.next().unwrap();
+    let reminder_event = iter.next().unwrap();
 
-    if let ChatEvent::Message(m) = latest_bot_message.event {
+    if let ChatEvent::Message(m) = reminder_created_event.event {
+        if let MessageContent::MessageReminderCreated(r) = m.content {
+            assert_eq!(r.notes, Some(notes.clone()));
+            assert!(r.hidden);
+        } else {
+            panic!()
+        }
+    } else {
+        panic!()
+    }
+
+    if let ChatEvent::Message(m) = reminder_event.event {
         if let MessageContent::MessageReminder(r) = m.content {
             assert_eq!(r.notes, Some(notes));
         } else {
@@ -116,9 +129,18 @@ fn cancel_message_reminder_succeeds() {
     env.advance_time(Duration::from_millis(1000));
     env.tick();
 
-    assert!(
-        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, latest_bot_message_index.incr(), true, 1, 1)
-            .events
-            .is_empty()
-    );
+    let events =
+        client::user::happy_path::events(env, &user1, OPENCHAT_BOT_USER_ID, latest_bot_message_index, true, 1000, 1000).events;
+
+    assert_eq!(events.len(), 1);
+
+    if let ChatEvent::Message(m) = &events[0].event {
+        if let MessageContent::MessageReminderCreated(r) = &m.content {
+            assert!(r.hidden);
+        } else {
+            panic!()
+        }
+    } else {
+        panic!()
+    }
 }